+++ packages/components/src/Button/ButtonSet.tsx
@@ -31,29 +31,18 @@ import React, {
   forwardRef,
   ForwardRefExoticComponent,
   isValidElement,
+  MouseEvent,
   ReactNode,
   Ref,
 } from 'react'
-import {
-  CompatibleHTMLProps,
-  BorderProps,
-  PositionProps,
-  SpaceProps,
-} from '@looker/design-tokens'
-import { Box } from '../Layout'
+import styled from 'styled-components'
+import { CompatibleHTMLProps } from '@looker/design-tokens'
+import { simpleLayoutCSS, SimpleLayoutProps } from '../Layout/utils/simple'
 import { ButtonItemProps } from './ButtonItem'
 
-export interface ButtonGroupOrToggleProps<
-  ValueType extends string | string[] = string[]
->
-  extends PositionProps,
-    BorderProps,
-    SpaceProps,
+interface ButtonSetProps<ValueType extends string | string[] = string[]>
+  extends SimpleLayoutProps,
     Omit<CompatibleHTMLProps<HTMLDivElement>, 'value' | 'onChange'> {
-  /**
-   * Internal use only
-   */
-  isControlled?: boolean
   /**
    * One or more ButtonItem
    */
@@ -62,29 +51,32 @@ export interface ButtonGroupOrToggleProps<
    * Value for controlling the component
    */
   value?: ValueType
+  onChange?: (e?: ChangeEvent<HTMLInputElement>) => void
+  isToggle?: boolean
   /**
-   * Change callback for controlling the component
+   * Value can be unset by clicking selected button (ButtonToggle only)
    */
-  onChange?: (value: ValueType) => void
+  nullable?: boolean
 }
 
-interface ButtonSetProps<ValueType extends string | string[] = string[]>
-  extends Omit<ButtonGroupOrToggleProps<ValueType>, 'onChange'> {
-  onChange?: (e: ChangeEvent<HTMLInputElement>) => void
-  isToggle?: boolean
+export interface ButtonGroupOrToggleBaseProps<
+  ValueType extends string | string[] = string[]
+> extends Omit<ButtonSetProps<ValueType>, 'isToggle' | 'onChange'> {
+  onChange?: (value: ValueType) => void
 }
 
 export type ButtonSetType<
   T extends string | string[] = string[]
 > = ForwardRefExoticComponent<ButtonSetProps<T> & { ref: Ref<HTMLDivElement> }>
 
-export const ButtonSet = forwardRef(
+export const ButtonSetLayout = forwardRef(
   (
     {
       children,
       disabled,
       isToggle,
       name,
+      nullable,
       onChange: groupOnChange,
       value,
       ...props
@@ -111,7 +103,9 @@ export const ButtonSet = forwardRef(
                 ? value === childValue
                 : value.includes(childValue),
             }
-          : { selected: childProps.selected }),
+          : typeof value === 'string' && value === ''
+          ? { selected: undefined }
+          : {}),
         ...(disabled ? { disabled } : {}),
         ...(groupOnChange
           ? {
@@ -121,22 +115,38 @@ export const ButtonSet = forwardRef(
               },
             }
           : {}),
+        ...(nullable && isToggle && groupOnChange
+          ? {
+              onClick: (e: MouseEvent<HTMLLabelElement>) => {
+                // The onClick is attached to the label but the browser will
+                // call it twice (2nd time with the input as target)
+                // If we un-check the radio in the 1st event (on label), the browser will immediately
+                // re - check it, so we wait for the 2nd event (on input)
+                if ((e.target as HTMLElement).tagName === 'INPUT') {
+                  groupOnChange()
+                }
+                childProps.onClick && childProps.onClick(e)
+              },
+            }
+          : {}),
       })
     })
 
     return (
-      <Box
-        alignItems="center"
-        display="inline-flex"
-        textAlign="center"
-        fontSize="small"
-        ref={ref}
-        {...props}
-      >
+      <div ref={ref} {...props}>
         {clonedChildren}
-      </Box>
+      </div>
     )
   }
 )
 
-ButtonSet.displayName = 'ButtonSet'
+ButtonSetLayout.displayName = 'ButtonSetLayout'
+
+export const ButtonSet = styled(ButtonSetLayout)`
+  ${simpleLayoutCSS}
+  align-items: center;
+  display: inline-flex;
+  flex-wrap: wrap;
+  font-size: ${({ theme }) => theme.fontSizes.small};
+  text-align: center;
+`